skaven_watch/requirements.txt
html5lib
lxml
requests
beautifulsoup4
discord.py
python-dotenv
pymongo
